Time: 881 ms (68.47%), Space: 298.2 MB (5.02%) - LeetHub

+++ 2316-count-unreachable-pairs-of-nodes-in-an-undirected-graph/2316-count-unreachable-pairs-of-nodes-in-an-undirected-graph.cpp
@@ -64,6 +64,6 @@ public:
         }
         if(components==1)return 0;
         
-        return ans/(long long int)2;
+        return ans/2ll;
     }
 };
